# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginUserDTO:
    properties:
      password:
        example: securepassword
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  auth.RegisterUserDTO:
    properties:
      password:
        example: securepassword
        minLength: 6
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  monitor.RegisterServiceDTO:
    properties:
      check_interval:
        example: 60
        minimum: 1
        type: integer
      name:
        example: My Service
        type: string
      url:
        example: https://example.com
        type: string
    required:
    - check_interval
    - name
    - url
    type: object
host: localhost:8080
info:
  contact: {}
  description: A distributed health monitoring system API.
  title: Health Checker API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Access token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /services:
    post:
      consumes:
      - application/json
      description: Register a new service for health monitoring
      parameters:
      - description: Service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/monitor.RegisterServiceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Service registered successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Register a new service
      tags:
      - services
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"